requirements.txt
streamlit
python-dotenv
youtube-transcript-api
langchain-community
langchain-text-splitters
langchain-core
langchain-huggingface
faiss-cpu
tiktoken
openai